エントリポイントのファイル名変更
