+++ 08.ls_object/ls.rb
@@ -3,7 +3,6 @@
 
 require 'optparse'
 require 'etc'
-require_relative 'lib/ls'
 require_relative 'lib/input_parser'
 require_relative 'lib/entries_generator'
 require_relative 'lib/entry'
@@ -12,4 +11,14 @@ require_relative 'lib/option_applier'
 require_relative 'lib/display_content_builder'
 require_relative 'lib/display'
 
-LS.run
+class Ls
+  def self.run
+    path, options = InputParser.parse.values_at(:path, :options)
+    entries = EntriesGenerator.new(path).generate
+    entries_after_options = OptionApplier.new(options, entries).apply
+    display_content = DisplayContentBuilder.build(entries_after_options)
+    Display.new(display_content, options).print
+  end
+end
+
+Ls.run
